Extract settings and landing view parts
