+++ src/app/views/settings-view.tsx
@@ -1,6 +1,4 @@
-import { Info, Search, X } from 'lucide-react'
-import { Fragment, useCallback, useEffect, useLayoutEffect, useRef, useState } from 'react'
-import { Tooltip } from '../components/common/tooltip'
+import { useCallback, useEffect, useLayoutEffect, useRef, useState } from 'react'
 import { ViewHeader } from '../components/common/view-header'
 import { ViewShell } from '../components/common/view-shell'
 import type {
@@ -13,33 +11,19 @@ import type {
   PiThemeState,
 } from '../desktop/types'
 import type { Project } from '../types'
-import { settingsSectionClass } from '../ui/classes'
 import { cn } from '../utils/cn'
-import { settingsHelpRowClass } from './settings/settingsClasses'
+import {
+  SettingsCategorySidebar,
+  SettingsGroupsList,
+  SettingsHeaderActions,
+  SettingsMobileFilters,
+  SettingsSearchField,
+} from './settings/settings-view-parts'
 import { buildSettingsDescriptors } from './settings/settingsDescriptors'
 import { normalizeManagedDictationModelId } from './settings/settingsDictationHelpers'
-import {
-  filterSettings,
-  groupSettingsByCategory,
-  settingsCategories,
-} from './settings/settingsGroups'
+import { filterSettings, groupSettingsByCategory } from './settings/settingsGroups'
 import type { SettingsCategoryId, SettingsOpenTarget } from './settings/settingsTypes'
-import { SettingRow } from './settings/settingsUi'
 import { useSettingsController } from './settings/useSettingsController'
-
-function DevBranchToggle({ checked, onToggle }: { checked: boolean; onToggle: () => void }) {
-  return (
-    <label className="mt-2 flex min-h-8 cursor-pointer items-center justify-between gap-2 px-3 text-[12px] text-[color:var(--muted)] transition-colors hover:text-[color:var(--text)]">
-      <span className="min-w-0 truncate">Dev branch</span>
-      <input
-        type="checkbox"
-        checked={checked}
-        onChange={onToggle}
-        className="h-3.5 w-3.5 shrink-0 accent-[color:var(--accent)]"
-      />
-    </label>
-  )
-}
 
 type SettingsViewProps = {
   appSettings: AppSettings
@@ -56,11 +40,6 @@ type SettingsViewProps = {
 
 function getDatasetValue(element: HTMLElement, key: string) {
   return element.dataset[key]
-}
-
-function getCategoryHelpIntro(categoryId: SettingsCategoryId) {
-  if (categoryId !== 'keybindings') return null
-  return 'Shortcuts require at least one modifier — Ctrl, Shift, Alt, or Command — plus one key.'
 }
 
 export function SettingsView({
@@ -337,57 +316,18 @@ export function SettingsView({
       <div className="grid min-w-0 items-center gap-4 lg:grid-cols-[220px_minmax(0,1fr)_auto]">
         <ViewHeader title="App settings" className="items-center" />
         <div className="hidden h-10 items-center lg:flex">
-          <label className="relative block w-[min(460px,42vw)]">
-            <Search
-              size={15}
-              className="pointer-events-none absolute top-1/2 left-3 -translate-y-1/2 text-[color:var(--muted)]"
-            />
-            <input
-              type="search"
-              value={filter}
-              onChange={(event) => setFilter(event.currentTarget.value)}
-              className="h-10 w-full min-w-0 flex-1 rounded-xl border border-[color:var(--border)] bg-[rgba(255,255,255,0.055)] px-3 py-2 pl-9 text-[13px] text-[color:var(--text)] outline-none placeholder:text-[color:var(--muted)]"
-              placeholder="Search…"
-              aria-label="Search settings"
-            />
-          </label>
+          <SettingsSearchField
+            value={filter}
+            onChange={setFilter}
+            className="w-[min(460px,42vw)]"
+          />
         </div>
-        <div className="flex items-center justify-end gap-2">
-          <Tooltip
-            content={
-              helpColumnAvailable
-                ? showHelp
-                  ? 'Hide setting descriptions'
-                  : 'Show setting descriptions'
-                : 'Window is too small for the help column. Hover settings to see tooltips instead.'
-            }
-            placement="left"
-          >
-            <button
-              type="button"
-              className={cn(
-                'inline-flex h-8 w-8 items-center justify-center self-center rounded-full border border-[color:var(--border)] bg-[rgba(255,255,255,0.03)] text-[color:var(--text)] transition-colors duration-150 ease-out hover:bg-[rgba(255,255,255,0.07)] disabled:cursor-not-allowed disabled:opacity-40',
-                showHelp && 'border-[color:var(--accent-border)] bg-[color:var(--accent-bg)]',
-              )}
-              onClick={() => setShowHelp((current) => !current)}
-              aria-label={showHelp ? 'Hide setting descriptions' : 'Show setting descriptions'}
-              aria-pressed={showHelp}
-              disabled={!helpColumnAvailable}
-            >
-              <Info size={14} />
-            </button>
-          </Tooltip>
-          <button
-            type="button"
-            className="inline-flex h-8 w-8 items-center justify-center self-center rounded-full border border-[color:var(--border)] bg-[rgba(255,255,255,0.03)] text-[color:var(--text)] transition-colors duration-150 ease-out hover:bg-[rgba(255,255,255,0.07)]"
-            onClick={closeSettings}
-            aria-label="Close app settings"
-            data-tooltip="Close app settings"
-            data-tooltip-placement="left"
-          >
-            <X size={14} />
-          </button>
-        </div>
+        <SettingsHeaderActions
+          helpColumnAvailable={helpColumnAvailable}
+          showHelp={showHelp}
+          onToggleHelp={() => setShowHelp((current) => !current)}
+          onClose={closeSettings}
+        />
       </div>
 
       <div
@@ -396,41 +336,13 @@ export function SettingsView({
           showHelp && 'lg:grid-cols-[220px_minmax(0,1fr)_minmax(18rem,24rem)]',
         )}
       >
-        <div className="sticky top-0 hidden max-h-full min-w-0 overflow-y-auto lg:grid">
-          <nav className="grid rounded-[22px] border border-[color:var(--border)] bg-[rgba(255,255,255,0.02)] p-2">
-            <button
-              type="button"
-              className={cn(
-                'flex h-10 items-center rounded-xl px-3 text-left text-[12px] transition-colors active:scale-[0.96]',
-                activeCategory === null && !normalizedFilter
-                  ? 'bg-[color:var(--accent-bg)] text-[color:var(--text)]'
-                  : 'text-[color:var(--muted)] hover:bg-[color:var(--surface-hover)] hover:text-[color:var(--text)]',
-              )}
-              onClick={() => setActiveCategory(null)}
-            >
-              All settings
-            </button>
-            {settingsCategories.map((category) => (
-              <button
-                key={category.id}
-                type="button"
-                className={cn(
-                  'flex h-10 items-center rounded-xl px-3 text-left text-[12px] transition-colors active:scale-[0.96]',
-                  activeCategory === category.id && !normalizedFilter
-                    ? 'bg-[color:var(--accent-bg)] text-[color:var(--text)]'
-                    : 'text-[color:var(--muted)] hover:bg-[color:var(--surface-hover)] hover:text-[color:var(--text)]',
-                )}
-                onClick={() => setActiveCategory(category.id)}
-              >
-                {category.label}
-              </button>
-            ))}
-          </nav>
-          <DevBranchToggle
-            checked={appSettings.devUpdateBranch}
-            onToggle={controller.toggleDevUpdateBranch}
-          />
-        </div>
+        <SettingsCategorySidebar
+          activeCategory={activeCategory}
+          normalizedFilter={normalizedFilter}
+          appSettings={appSettings}
+          onSelectCategory={setActiveCategory}
+          onToggleDevBranch={controller.toggleDevUpdateBranch}
+        />
 
         <div
           ref={settingsScrollRef}
@@ -439,125 +351,21 @@ export function SettingsView({
             showHelp && 'lg:col-span-2 lg:grid-cols-[minmax(0,1fr)_minmax(18rem,24rem)] lg:gap-x-4',
           )}
         >
-          <div className="grid min-w-0 content-start gap-4 lg:hidden">
-            <label className="relative block lg:hidden">
-              <Search
-                size={15}
-                className="pointer-events-none absolute top-1/2 left-3 -translate-y-1/2 text-[color:var(--muted)]"
-              />
-              <input
-                type="search"
-                value={filter}
-                onChange={(event) => setFilter(event.currentTarget.value)}
-                className="h-10 w-full min-w-0 flex-1 rounded-xl border border-[color:var(--border)] bg-[rgba(255,255,255,0.055)] px-3 py-2 pl-9 text-[13px] text-[color:var(--text)] outline-none placeholder:text-[color:var(--muted)]"
-                placeholder="Search…"
-                aria-label="Search settings"
-              />
-            </label>
+          <SettingsMobileFilters
+            filter={filter}
+            activeCategory={activeCategory}
+            appSettings={appSettings}
+            onFilterChange={setFilter}
+            onSelectCategory={setActiveCategory}
+            onToggleDevBranch={controller.toggleDevUpdateBranch}
+          />
 
-            <div className="flex flex-wrap items-center gap-1.5 lg:hidden">
-              <button
-                type="button"
-                className={cn(
-                  'rounded-full border border-[color:var(--border)] px-3 py-1.5 text-[12px] transition-colors',
-                  activeCategory === null && 'bg-[color:var(--accent-bg)] text-[color:var(--text)]',
-                )}
-                onClick={() => setActiveCategory(null)}
-              >
-                All
-              </button>
-              {settingsCategories.map((category) => (
-                <button
-                  key={category.id}
-                  type="button"
-                  className={cn(
-                    'rounded-full border border-[color:var(--border)] px-3 py-1.5 text-[12px] text-[color:var(--muted)] transition-colors',
-                    activeCategory === category.id &&
-                      'bg-[color:var(--accent-bg)] text-[color:var(--text)]',
-                  )}
-                  onClick={() => setActiveCategory(category.id)}
-                >
-                  {category.label}
-                </button>
-              ))}
-              <DevBranchToggle
-                checked={appSettings.devUpdateBranch}
-                onToggle={controller.toggleDevUpdateBranch}
-              />
-            </div>
-          </div>
-
-          {visibleGroups.length > 0 ? (
-            visibleGroups.map((group) => (
-              <Fragment key={group.id}>
-                <section
-                  className={cn(
-                    settingsSectionClass,
-                    'motion-surface-pulse motion-settings-section-pulse min-w-0 gap-1 p-2.5',
-                  )}
-                  data-pulse-active={group.id === highlightedCategoryId ? 'true' : 'false'}
-                >
-                  <div
-                    className={cn(
-                      'flex items-baseline justify-between gap-3 px-1 pt-1 pb-1',
-                      getCategoryHelpIntro(group.id) && 'min-h-10 items-start',
-                    )}
-                  >
-                    <h2 className="text-[15px] font-semibold text-[color:var(--text)]">
-                      {group.label}
-                    </h2>
-                  </div>
-                  <div className="grid">
-                    {group.settings.map((setting) => (
-                      <SettingRow key={setting.id} setting={setting} showHelp={showHelp} />
-                    ))}
-                  </div>
-                </section>
-                {showHelp ? (
-                  <aside className="hidden min-w-0 content-start gap-1 rounded-[18px] border border-transparent p-2.5 lg:grid">
-                    <div
-                      className={cn(
-                        'flex items-baseline gap-3 px-1 pt-1 pb-1',
-                        getCategoryHelpIntro(group.id) && 'min-h-10 items-start',
-                      )}
-                    >
-                      {getCategoryHelpIntro(group.id) ? (
-                        <span className="min-w-0 text-[11.5px] leading-4 text-wrap text-[color:var(--muted)]">
-                          {getCategoryHelpIntro(group.id)}
-                        </span>
-                      ) : (
-                        <h2 className="invisible text-[15px] font-semibold">{group.label}</h2>
-                      )}
-                    </div>
-                    <div className="grid">
-                      {group.settings.map((setting) => (
-                        <div
-                          key={setting.id}
-                          className={settingsHelpRowClass}
-                          style={
-                            settingRowHeights[setting.id]
-                              ? { height: `${settingRowHeights[setting.id]}px` }
-                              : undefined
-                          }
-                        >
-                          <span className="relative top-[10px] truncate">
-                            {setting.helpDescription ?? setting.description}
-                          </span>
-                        </div>
-                      ))}
-                    </div>
-                  </aside>
-                ) : null}
-              </Fragment>
-            ))
-          ) : (
-            <div className="rounded-[22px] border border-[rgba(169,178,215,0.12)] bg-[rgba(255,255,255,0.025)] p-8 text-center lg:col-span-full">
-              <div className="text-[14px] text-[color:var(--text)]">No matching settings</div>
-              <div className="mt-1 text-[12px] text-[color:var(--muted)]">
-                Try a broader term like “Pi”, “model”, “folder”, or “voice”.
-              </div>
-            </div>
-          )}
+          <SettingsGroupsList
+            visibleGroups={visibleGroups}
+            showHelp={showHelp}
+            highlightedCategoryId={highlightedCategoryId}
+            settingRowHeights={settingRowHeights}
+          />
         </div>
       </div>
     </ViewShell>
